[REFACTOR] Remove namespacing from vocabulary Terms
**ISSUE**
Namspacing Terms under Vocabularies adds complexity to controllers and
views without significant benefit. While the number of top-level
models is growing and a term is only meaningful within the context of
a parent vocabulary, we haven't implemented namspacing on any other
models. Namepacing Terms adds to the congnitive load and complexity
of the code without providing any end-user or maintenance value.

**FIX**
Re-arrange the models to work without namespacing.
