+++ db/migrate/20240628173658_rename_vocabulary_terms_to_terms.rb
@@ -1,0 +1,5 @@
+class RenameVocabularyTermsToTerms < ActiveRecord::Migration[7.1]
+  def change
+    rename_table 'vocabulary_terms', 'terms'
+  end
+end
