--- spec/factories/vocabulary/terms.rb
@@ -1,6 +1,0 @@
-FactoryBot.define do
-  factory :vocabulary_term, class: 'Vocabulary::Term' do
-    vocabulary { nil }
-    label { Faker::Lorem.unique.words.join(' ').titleize }
-  end
-end
